Clean-MODG/requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
PyYAML
tqdm
Pillow
opencv-python
matplotlib
albumentations
shapely
